completed

+++ src/main/java/com/jobfinder/springboot/assignmentdemo/controller/JobController.java
@@ -4,13 +4,17 @@ package com.jobfinder.springboot.assignmentdemo.controller;
 import com.jobfinder.springboot.assignmentdemo.entity.Job;
 
 
+
+import com.jobfinder.springboot.assignmentdemo.exception.JobNotFoundException;
 import com.jobfinder.springboot.assignmentdemo.service.JobService;
 import com.jobfinder.springboot.assignmentdemo.service.LocationService;
 import com.jobfinder.springboot.assignmentdemo.service.SkillService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
-import javax.websocket.server.PathParam;
+
+import java.util.Arrays;
 import java.util.List;
 
 @RestController
@@ -61,7 +65,7 @@ public class JobController {
     public Job getJob(@PathVariable int jobId){
         Job theJob=jobService.findJobById(jobId);
         if(theJob==null){
-            throw new RuntimeException("Job id is not found");
+            throw new JobNotFoundException("Job id is not found");
 
         }
         return theJob;
@@ -71,46 +75,77 @@ public class JobController {
     //Delete job
     @DeleteMapping("jobs/{jobId}")
     public String deleteJob(@PathVariable int jobId){
+
+
+
+        Job job=jobService.findJobById(jobId);
+
+        if(job==null){
+            throw new JobNotFoundException("No job found with id "+jobId);
+        }
         jobService.deleteJob(jobId);
-        return "deleted job with id"+jobId;
+        return "Deleted job with id "+jobId;
     }
 
+    public void isAllStrings(String strings[],String type) throws Exception {
+        for(int i=0;i<strings.length;i++){
+            String location = strings[i];
+            boolean isNumeric = location.chars().allMatch( Character::isDigit );
+            if(isNumeric){
+                throw new Exception("Please provide valid "+type);
+            }
+        }
+
+    }
 
     @GetMapping("/jobs/filter")
-    public List<Job> filter(@RequestParam(required=false) String skills[],@RequestParam(required=false) String locations[]){
+    public List<Job> filter(@RequestParam(required=false) String locations[],@RequestParam(required=false) String skills[]) throws Exception {
+        List<Job> jobs;
 
-        System.out.println(skills[0]);
-        System.out.println(locations[0]);
-
-
-
+        Boolean skillsGiven=false;
+        Boolean locationsGiven=false;
 
 
-        List<Integer>  ids=skillService.getIdsWithSkills(skills);
-        System.out.println(ids);
+        if(skills==null || skills.length==0){
 
-        //List<Job> theJob=jobService.getJobsByIds(ids);
-        List<Job> theJob=jobService.filter(skills,locations);
-        if(theJob==null){
-            throw new RuntimeException("No Jobs is not found");
+            isAllStrings(locations,"locations");
+            skillsGiven=false;
+            List<Integer>  ids=locationService.getJobIdsInLocation(locations);
+            jobs=jobService.getJobsByIds(ids);
+        }
+        else if(locations==null || locations.length==0){
+            isAllStrings(skills,"skills");
+            locationsGiven=false;
+            List<Integer> ids=skillService.getIdsWithSkills(skills);
+            jobs=jobService.getJobsByIds(ids);
+        }
+        else {
+            isAllStrings(locations,"locations");
+            isAllStrings(skills,"skills");
+            List<Integer>  idsInLocation=locationService.getJobIdsInLocation(locations);
+            List<Integer> idsWithSkills=skillService.getIdsWithSkills(skills);
+            idsWithSkills.retainAll(idsInLocation);
+            jobs=jobService.getJobsByIds(idsWithSkills);
+        }
+
+        if(jobs==null || jobs.isEmpty()){
+            if(locationsGiven){
+                throw new JobNotFoundException("Currently No Jobs available In this locations "+Arrays.toString(locations));
+            }
+            else if(skillsGiven){
+                throw new JobNotFoundException("Currently No Jobs available with these skills "+ Arrays.toString(skills));
+            }
+            else {
+                throw new JobNotFoundException("Currently No Jobs available In this locations "+Arrays.toString(locations)+" with these skills"+Arrays.toString(skills));
+            }
+
 
         }
-        return theJob;
+        return jobs;
     }
 
 
 
-
-
-
-
-
-    //Delete job
-//    @DeleteMapping("jobs/{jobId}")
-//    public String deleteJob(@PathVariable int jobId){
-//        locationService.deleteSkill(jobId);
-//        return "deleted job with id"+jobId;
-//    }
